+++ packages/blocks/src/page-block/edgeless/edgeless-page-block.ts
@@ -11,7 +11,7 @@ import {
   BLOCK_ID_ATTR,
   hotkey,
   HOTKEYS,
-  resetNativeSeletion,
+  resetNativeSelection,
 } from '../../__internal__';
 import {
   EdgelessSelectionManager,
@@ -20,6 +20,12 @@ import {
   XYWH,
 } from './selection-manager';
 import style from './style.css';
+import {
+  bindCommonHotkey,
+  handleBackspace,
+  removeCommonHotKey,
+  tryUpdateGroupSize,
+} from '../utils';
 
 export interface EdgelessContainer extends HTMLElement {
   readonly store: Store;
@@ -72,14 +78,19 @@ export class EdgelessPageBlockComponent
 
   private _bindHotkeys() {
     const { store } = this;
-    hotkey.addListener(HOTKEYS.UNDO, () => store.undo());
-    hotkey.addListener(HOTKEYS.REDO, () => store.redo());
+    hotkey.addListener(HOTKEYS.BACKSPACE, this._backspace.bind(this));
+    bindCommonHotkey(store);
   }
 
   private _removeHotkeys() {
-    hotkey.removeListener([HOTKEYS.UNDO, HOTKEYS.REDO]);
+    hotkey.removeListener([HOTKEYS.BACKSPACE]);
+    removeCommonHotKey();
   }
-
+  _backspace(e: KeyboardEvent) {
+    if (this._selection.state.type === 'single') {
+      handleBackspace(this.store, e);
+    }
+  }
   private _initViewport() {
     const bound = this.mouseRoot.getBoundingClientRect();
     this.viewport.setSize(bound.width, bound.height);
@@ -92,7 +103,7 @@ export class EdgelessPageBlockComponent
   private _clearSelection() {
     requestAnimationFrame(() => {
       if (!this._selection.isActive) {
-        resetNativeSeletion(null);
+        resetNativeSelection(null);
       }
     });
   }
@@ -105,7 +116,6 @@ export class EdgelessPageBlockComponent
   update(changedProperties: Map<string, unknown>) {
     if (changedProperties.has('mouseRoot') && changedProperties.has('store')) {
       this._selection = new EdgelessSelectionManager(this);
-      this._initViewport();
     }
     super.update(changedProperties);
   }
@@ -124,6 +134,15 @@ export class EdgelessPageBlockComponent
 
     this._bindHotkeys();
 
+    this.addEventListener('keydown', e => {
+      if (e.ctrlKey || e.metaKey || e.shiftKey) return;
+      tryUpdateGroupSize(this.store, this.viewport.zoom);
+    });
+
+    requestAnimationFrame(() => {
+      this._initViewport();
+      this.requestUpdate();
+    });
     // XXX: should be called after rich text components are mounted
     this._clearSelection();
   }
@@ -145,7 +164,8 @@ export class EdgelessPageBlockComponent
       this.viewport
     );
 
-    const selectedRect = EdgelessSelectedRect(this._selection.state);
+    const { zoom } = this.viewport;
+    const selectedRect = EdgelessSelectedRect(this._selection.state, zoom);
 
     return html`
       <style></style>