+++ packages/editor/src/components/editor-container/editor-container.ts
@@ -1,5 +1,5 @@
-import { LitElement, html } from 'lit';
-import { customElement, state, query } from 'lit/decorators.js';
+import { html, LitElement } from 'lit';
+import { customElement, property, query, state } from 'lit/decorators.js';
 import { choose } from 'lit/directives/choose.js';
 
 import { Store } from '@blocksuite/store';
@@ -8,20 +8,10 @@ import { BlockSchema } from '../../block-loader';
 
 type PageBlockModel = InstanceType<typeof BlockSchema.page>;
 
-const params = new URLSearchParams(location.search);
-const room = params.get('room') || 'virgo-default';
-
-const IS_PLAYGROUND = location.href.includes('5173');
-
-const options = {
-  room,
-  useDebugProvider: IS_PLAYGROUND,
-};
-
 @customElement('editor-container')
 export class EditorContainer extends LitElement {
-  @state()
-  store = new Store(options).register(BlockSchema);
+  @property()
+  store!: Store;
 
   @state()
   mode: 'page' | 'edgeless' = 'page';
@@ -43,30 +33,31 @@ export class EditorContainer extends LitElement {
   private _placeholderInput!: HTMLInputElement;
 
   @state()
-  placeholderModel = new BlockSchema.page(this.store, {});
+  placeholderModel!: PageBlockModel;
 
-  constructor() {
-    super();
+  @state()
+  unsubscribe = [] as (() => void)[];
 
-    this._subscribeStore();
-
-    // @ts-ignore
-    window.store = this.store;
-    // @ts-ignore
-    window.editor = this;
+  update(changedProperties: Map<string, unknown>) {
+    if (changedProperties.has('store')) {
+      this.placeholderModel = new BlockSchema.page(this.store, {});
+    }
+    super.update(changedProperties);
   }
 
   private _subscribeStore() {
     // if undo to empty page, reset to empty placeholder
-    this.store.signals.updated.on(() => {
+    const unsubscribeUpdate = this.store.signals.updated.on(() => {
       this.isEmptyPage = this.store.isEmpty;
     });
+    this.unsubscribe.push(unsubscribeUpdate.dispose);
 
-    this.store.signals.rootAdded.on(block => {
+    const unsubscribeRootAdd = this.store.signals.rootAdded.on(block => {
       this.model = block as PageBlockModel;
       this.model.childrenUpdated.on(() => this.requestUpdate());
       this.requestUpdate();
     });
+    this.unsubscribe.push(unsubscribeRootAdd.dispose);
   }
 
   // disable shadow DOM to workaround quill
@@ -79,7 +70,17 @@ export class EditorContainer extends LitElement {
       this.mode = detail;
     });
 
+    this._subscribeStore();
+
     this._placeholderInput?.focus();
+    // @ts-ignore
+    window.store = this.store;
+    // @ts-ignore
+    window.editor = this;
+  }
+
+  disconnectedCallback() {
+    this.unsubscribe.forEach(fn => fn());
   }
 
   render() {